Revert PageHeaderImage abstraction
