+++ apps/ffdweb-site/src/app/_components/PageHeader/PageHeaderImage.tsx
@@ -1,25 +1,39 @@
-import Image from 'next/image'
+import Image, { type ImageProps } from 'next/image'
 
 import { clsx } from 'clsx'
 
-import {
-  getPageHeaderImageProps,
-  type PageHeaderImageProps,
-} from '@filecoin-foundation/ui/PageHeader'
+import { buildImageSizeProp } from '@filecoin-foundation/utils/buildImageSizeProp'
+import type {
+  ImageObjectFit,
+  StaticImageProps,
+} from '@filecoin-foundation/utils/types/imageType'
+
+type PageHeaderImageProps = (StaticImageProps | ImageProps) & {
+  objectFit?: ImageObjectFit
+}
 
 export function PageHeaderImage(image: PageHeaderImageProps) {
   const isStaticImage = 'data' in image
-  const commonProps = getPageHeaderImageProps(image, {
-    startSize: '100vw',
-    lg: '490px',
-  })
+
+  const commonProps = {
+    alt: image.alt,
+    priority: true,
+    quality: 100,
+    sizes: buildImageSizeProp({ startSize: '390px' }),
+    className: clsx(
+      'page-header-image',
+      image.objectFit?.objectFit === 'cover' && 'object-cover',
+      image.objectFit?.objectFit === 'contain' && 'object-contain',
+    ),
+  }
 
   if (isStaticImage) {
     return (
       <Image
         {...commonProps}
-        className={clsx(commonProps.className, 'aspect-video')}
+        className={clsx(commonProps.className, 'aspect-square w-full max-w-96')}
         src={image.data}
+        alt={commonProps.alt}
       />
     )
   }
