+++ packages/ui/src/PageHeader/PageHeader.tsx
@@ -5,8 +5,7 @@ import {
 import { Meta, type MetaDataType } from '@filecoin-foundation/ui/Meta'
 import { TagLabel } from '@filecoin-foundation/ui/TagComponents'
 
-import { PageHeaderImage } from './PageHeaderImage'
-import type { PageHeaderImageProps } from './pageHeaderImageHelpers'
+import { PageHeaderImage, type PageHeaderImageProps } from './PageHeaderImage'
 import { PageHeaderTitle, type PageHeaderTitleProps } from './PageHeaderTitle'
 
 type PageHeaderDescriptionProps = {
