+++ packages/ui/src/PageHeader/index.ts
@@ -1,4 +1,3 @@
 export * from './PageHeader'
 export * from './PageHeaderImage'
-export * from './pageHeaderImageHelpers'
 export * from './PageHeaderTitle'
